# examples/server/server.ts
import { createServer, IncomingMessage, ServerResponse } from "node:http";
import { randomUUID } from "node:crypto";
import { readFileSync } from "node:fs";
import {
  createVerifierHandlers,
  OpenEudiEngine,
  Openid4vpEngine,
  MemoryKVStore,
  clientIpFromHeaders,
  type RequestContext,
  type HandlerResponse,
  type VerifiedClaims,
  type VerifierEngine,
  type VerifierMode,
} from "@eudi-verify/server";

const PORT = parseInt(process.env.PORT || "3000", 10);
const HOST = process.env.HOST || "127.0.0.1";
const BASE_URL = process.env.BASE_URL || `http://localhost:${PORT}/api/eudi`;
const DEMO_RECEIPT_TTL_MS = 5 * 60 * 1000;

/** `EUDI_MODE=production` → Openid4vpEngine; default remains demo OpenEudiEngine. */
const MODE: VerifierMode =
  process.env.EUDI_MODE === "production" ? "production" : "demo";

function buildEngine(): VerifierEngine {
  if (MODE !== "production") {
    return new OpenEudiEngine({
      mode: "demo",
      baseUrl: BASE_URL,
      demoDelayMs: 1500,
    });
  }

  const trustMode = process.env.EUDI_TRUST || "skip"; // skip|static
  const trustedCertsPath = process.env.EUDI_TRUSTED_CERTS;
  const trustConfig =
    trustMode === "static" && trustedCertsPath
      ? { trustedCerts: [new Uint8Array(readFileSync(trustedCertsPath))] }
      : ({
          skipTrustCheck: true,
          acknowledgeInsecureTrust: true,
        } as const);

  return new Openid4vpEngine({
    mode: "production",
    baseUrl: BASE_URL,
    allowInsecureTransport: !BASE_URL.startsWith("https://"),
    ...trustConfig,
  });
}

interface DemoReceipt {
  sessionId: string;
  claims: VerifiedClaims;
  verifiedAt: string;
  token: string;
}

interface DemoReceiptDTO {
  sessionId: string;
  claims: VerifiedClaims;
  verifiedAt: string;
  tokenPreview: string;
  tokenFormat: string;
}

const engine = buildEngine();
const store = new MemoryKVStore();
const handlers = createVerifierHandlers({
  engine,
  store,
  baseUrl: BASE_URL,
  mode: MODE,
  tokenSecret: process.env.TOKEN_SECRET || "demo-secret-change-in-production",
  rateLimit: { maxRequests: 10, windowMs: 60_000 },
});

function modeHeader(): Record<string, string> {
  return { "X-Eudi-Mode": MODE };
}

function demoReceiptKey(rid: string): string {
  return `demo-receipt:${rid}`;
}

function truncateToken(token: string): string {
  if (token.length <= 24) return token;
  return `${token.slice(0, 12)}…${token.slice(-8)}`;
}

// ponytail: read sid from verified token payload; safe only after verifyToken succeeds
function sessionIdFromToken(token: string): string | undefined {
  const parts = token.split(".");
  if (parts.length !== 3) return undefined;
  try {
    const payload = JSON.parse(Buffer.from(parts[1], "base64url").toString());
    return typeof payload.sid === "string" && payload.sid
      ? payload.sid
      : undefined;
  } catch {
    return undefined;
  }
}

function receiptToDTO(receipt: DemoReceipt): DemoReceiptDTO {
  return {
    sessionId: receipt.sessionId,
    claims: receipt.claims,
    verifiedAt: receipt.verifiedAt,
    tokenPreview: truncateToken(receipt.token),
    tokenFormat: "eudi_v1.<payload>.<hmac>",
  };
}

async function parseBody(req: IncomingMessage): Promise<string> {
  const chunks: Buffer[] = [];
  for await (const chunk of req) chunks.push(chunk as Buffer);
  return Buffer.concat(chunks).toString();
}

function buildContext(
  req: IncomingMessage,
  params: Record<string, string>,
  body?: unknown,
  rawBody?: string,
): RequestContext {
  return {
    ip: clientIpFromHeaders(req.headers, req.socket.remoteAddress),
    origin: req.headers.origin,
    params,
    body,
    rawBody,
  };
}

async function sendResponse(
  res: ServerResponse,
  result: HandlerResponse,
): Promise<void> {
  const headers: Record<string, string> = {
    "Content-Type": "application/json",
    ...result.headers,
  };
  res.writeHead(result.status, headers);
  res.end(
    typeof result.body === "string" ? result.body : JSON.stringify(result.body),
  );
}

const server = createServer(async (req, res) => {
  const url = new URL(req.url || "/", `http://localhost:${PORT}`);
  const path = url.pathname;

  // CORS for local dev (frontend on different port)
  if (req.method === "OPTIONS") {
    res.writeHead(204, {
      "Access-Control-Allow-Origin": req.headers.origin || "*",
      "Access-Control-Allow-Methods": "GET, POST, OPTIONS",
      "Access-Control-Allow-Headers": "Content-Type",
      "Access-Control-Max-Age": "86400",
    });
    return res.end();
  }
  res.setHeader("Access-Control-Allow-Origin", req.headers.origin || "*");

  // Demo-only API routes
  if (path.startsWith("/api/demo")) {
    const demoPath = path.replace("/api/demo", "");

    const receiptMatch = demoPath.match(/^\/receipt\/([^/]+)$/);
    if (req.method === "GET" && receiptMatch) {
      const rid = decodeURIComponent(receiptMatch[1]);
      const receipt = await store.get<DemoReceipt>(demoReceiptKey(rid));
      if (!receipt) {
        res.writeHead(404, { "Content-Type": "application/json" });
        return res.end(
          JSON.stringify({
            error: "not_found",
            message: "Receipt not found or expired",
          }),
        );
      }
      res.writeHead(200, {
        "Content-Type": "application/json",
        ...modeHeader(),
      });
      return res.end(JSON.stringify(receiptToDTO(receipt)));
    }

    if (req.method === "POST" && demoPath === "/replay") {
      const raw = await parseBody(req);
      let body: { rid?: string };
      try {
        body = JSON.parse(raw);
      } catch {
        res.writeHead(400, { "Content-Type": "application/json" });
        return res.end(
          JSON.stringify({
            error: "bad_request",
            message: "Invalid JSON body",
          }),
        );
      }

      const rid = body.rid;
      if (!rid) {
        res.writeHead(400, { "Content-Type": "application/json" });
        return res.end(
          JSON.stringify({ error: "bad_request", message: "Missing rid" }),
        );
      }

      const receipt = await store.get<DemoReceipt>(demoReceiptKey(rid));
      if (!receipt) {
        res.writeHead(404, { "Content-Type": "application/json" });
        return res.end(
          JSON.stringify({
            error: "not_found",
            message: "Receipt not found or expired",
          }),
        );
      }

      const result = await handlers.verifyToken(
        buildContext(
          req,
          {},
          { token: receipt.token },
          JSON.stringify({ token: receipt.token }),
        ),
      );
      res.writeHead(200, {
        "Content-Type": "application/json",
        ...modeHeader(),
      });
      return res.end(JSON.stringify(result.body));
    }

    res.writeHead(404, { "Content-Type": "application/json" });
    return res.end(
      JSON.stringify({
        error: "not_found",
        message: "Demo API endpoint not found",
      }),
    );
  }

  // API routes
  if (path.startsWith("/api/eudi")) {
    const apiPath = path.replace("/api/eudi", "");

    if (apiPath === "/sessions") {
      if (req.method === "HEAD") {
        res.writeHead(200, {
          "Content-Type": "application/json",
          ...modeHeader(),
        });
        return res.end();
      }
      if (req.method === "POST") {
        const raw = await parseBody(req);
        const body = JSON.parse(raw);
        return sendResponse(
          res,
          await handlers.createSession(buildContext(req, {}, body, raw)),
        );
      }
    }

    const cancelMatch = apiPath.match(/^\/sessions\/([^/]+)\/cancel$/);
    if (cancelMatch && req.method === "POST") {
      const [, sessionId] = cancelMatch;
      return sendResponse(
        res,
        await handlers.cancelSession(buildContext(req, { sessionId })),
      );
    }

    const sessionMatch = apiPath.match(/^\/sessions\/([^/]+)$/);
    if (sessionMatch) {
      const [, sessionId] = sessionMatch;
      if (req.method === "GET") {
        return sendResponse(
          res,
          await handlers.getSession(buildContext(req, { sessionId })),
        );
      }
    }

    if (req.method === "POST" && apiPath === "/tokens/verify") {
      const raw = await parseBody(req);
      const body = JSON.parse(raw);
      return sendResponse(
        res,
        await handlers.verifyToken(buildContext(req, {}, body, raw)),
      );
    }

    if (req.method === "POST" && apiPath === "/callback") {
      const raw = await parseBody(req);
      return sendResponse(
        res,
        await handlers.handleCallback(buildContext(req, {}, undefined, raw)),
      );
    }

    const requestMatch = apiPath.match(/^\/request\/([^/]+)$/);
    if (requestMatch && req.method === "GET") {
      return sendResponse(
        res,
        await handlers.getRequest(
          buildContext(req, { requestId: requestMatch[1] }),
        ),
      );
    }

    res.writeHead(404, { "Content-Type": "application/json" });
    return res.end(
      JSON.stringify({ error: "not_found", message: "API endpoint not found" }),
    );
  }

  if (req.method === "POST" && path === "/api/checkout") {
    const raw = await parseBody(req);
    const params = new URLSearchParams(raw);
    const token = params.get("eudi_token");
    const sessionId = params.get("eudi_session_id") || "";

    if (!token) {
      res.writeHead(400, { "Content-Type": "application/json" });
      return res.end(JSON.stringify({ error: "missing_token" }));
    }

    const result = await handlers.verifyToken(
      buildContext(req, {}, { token }, raw),
    );
    const verifyResult = result.body as {
      valid: boolean;
      claims?: VerifiedClaims;
      sessionId?: string;
    };

    if (verifyResult.valid && verifyResult.claims) {
      const rid = randomUUID();
      const receipt: DemoReceipt = {
        sessionId:
          verifyResult.sessionId ||
          sessionId ||
          sessionIdFromToken(token) ||
          "",
        claims: verifyResult.claims,
        verifiedAt: new Date().toISOString(),
        token,
      };
      await store.set(demoReceiptKey(rid), receipt, DEMO_RECEIPT_TTL_MS);
      res.writeHead(200, { "Content-Type": "application/json" });
      return res.end(JSON.stringify({ success: true, receiptId: rid }));
    }

    res.writeHead(400, { "Content-Type": "application/json" });
    return res.end(JSON.stringify({ error: "invalid_token" }));
  }

  res.writeHead(404, { "Content-Type": "application/json" });
  res.end(
    JSON.stringify({ error: "not_found", message: "Endpoint not found" }),
  );
});

server.listen(PORT, HOST, () => {
  console.log(
    `\n🔵 EUDI Verify API Server running at http://${HOST}:${PORT}\n`,
  );
  console.log("Frontend examples should proxy /api/* to this server.\n");
  if (MODE === "demo") {
    console.warn(
      "[!] DEMO MODE - Credentials are simulated. Do not use in production.\n",
    );
  } else {
    const trust = process.env.EUDI_TRUST || "skip";
    console.log(`[+] PRODUCTION MODE (Openid4vpEngine), EUDI_TRUST=${trust}`);
    console.log(`    BASE_URL=${BASE_URL}`);
    if (trust === "skip") {
      console.warn(
        "[!] EUDI_TRUST=skip — issuer trust anchoring DISABLED (lab-only).\n",
      );
    }
  }
});